🏷️ Get AcceptedPlugin type dynamically
In postcss@<5.2.15, the arguments type of postcss() function was defined with the name "acceptedPlugin".

https://github.com/postcss/postcss/blob/5.2.14/lib/postcss.d.ts#L9-L11

For this reason, the name "AcceptedPlugin" defined in the latest postcss@7.0.18 does not work with the old postcss included type definition.
In order to solve this problem, the "AcceptedPlugin" type is obtained by using Utility Type: Parameters<T>.