✅  Add symbolic links of node_modules directory
Metalsmith CLI only import plugins from the node_modules directory directly under the current working directory.
When loading a locall installed package, the node_modules directory must be in the same hierarchy as the metalsmith.json file.
To solve this problem, I added a symbolic link that points to the metalsmith-postcss2 directory.